Se agrego badRequestException en el controlador de posts.

+++ src/main/java/com/fiuni/adoptamena/api/controller/post/PostController.java
@@ -2,12 +2,14 @@ package com.fiuni.adoptamena.api.controller.post;
 
 import com.fiuni.adoptamena.api.dto.post.PostDTO;
 import com.fiuni.adoptamena.api.service.post.IPostService;
+import com.fiuni.adoptamena.exception_handler.exceptions.BadRequestException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -23,7 +25,12 @@ public class PostController {
     private IPostService postService;
 
     @GetMapping("/{id}")
-    public ResponseEntity<PostDTO> getPostById(@PathVariable(name = "id") int id) {
+    public ResponseEntity<PostDTO> getPostById(@PathVariable(name = "id") int id, BindingResult bindingResult) {
+
+        if (bindingResult.hasErrors()) {
+            throw new BadRequestException(bindingResult.getAllErrors());
+        }
+
         PostDTO data = this.postService.getById(id);
         return ResponseEntity.status(HttpStatus.OK).body(data);
     }
@@ -36,8 +43,13 @@ public class PostController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "content", required = false) String content,
             @RequestParam(value = "user", required = false) Integer userId, // Filtro por usuario
-            @RequestParam(value = "postType", required = false) Integer postTypeId // Filtro por tipo de post
+            @RequestParam(value = "postType", required = false) Integer postTypeId, // Filtro por tipo de post
+            BindingResult bindingResult
     ) {
+
+        if (bindingResult.hasErrors()) {
+            throw new BadRequestException(bindingResult.getAllErrors());
+        }
 
         String[] sortParams = sort.split(",");
         Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Order.asc(sortParams[0])));
@@ -48,7 +60,11 @@ public class PostController {
     }
 
     @PostMapping({ "", "/" })
-    public ResponseEntity<PostDTO> create(@RequestBody() PostDTO postDto) {
+    public ResponseEntity<PostDTO> create(@RequestBody() PostDTO postDto, BindingResult bindingResult) {
+
+        if (bindingResult.hasErrors()) {
+            throw new BadRequestException(bindingResult.getAllErrors());
+        }
 
         PostDTO data = this.postService.create(postDto);
         return ResponseEntity.status(HttpStatus.CREATED).body(data);
@@ -56,14 +72,24 @@ public class PostController {
 
     @PutMapping({ "/{id}" })
     public ResponseEntity<PostDTO> update(@PathVariable(name = "id", required = true) Integer id,
-            @RequestBody() PostDTO postDto) {
+            @RequestBody() PostDTO postDto, BindingResult bindingResult) {
+
+        if (bindingResult.hasErrors()) {
+            throw new BadRequestException(bindingResult.getAllErrors());
+        }
+
         postDto.setId(id);
         PostDTO data = this.postService.update(postDto);
         return ResponseEntity.status(HttpStatus.OK).body(data);
     }
 
     @DeleteMapping({ "/{id}" })
-    public ResponseEntity<String> delete(@PathVariable(name = "id", required = true) Integer id) {
+    public ResponseEntity<String> delete(@PathVariable(name = "id", required = true) Integer id, BindingResult bindingResult) {
+
+        if (bindingResult.hasErrors()) {
+            throw new BadRequestException(bindingResult.getAllErrors());
+        }
+
         this.postService.delete(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).body("post with id: " + id + "was deleted");
     }
@@ -73,7 +99,13 @@ public class PostController {
             @RequestParam(value = "page", defaultValue = "0") int page,
             @RequestParam(value = "size", defaultValue = "10") int size,
             @RequestParam(value = "sort", defaultValue = "id,asc") String sort,
-            @RequestParam(value = "keyword") String keyword) {
+            @RequestParam(value = "keyword") String keyword,
+            BindingResult bindingResult) {
+
+        if (bindingResult.hasErrors()) {
+            throw new BadRequestException(bindingResult.getAllErrors());
+        }
+
         String[] sortParams = sort.split(",");
         Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Order.asc(sortParams[0])));
 
@@ -83,7 +115,12 @@ public class PostController {
     }
 
     @PutMapping("/{id}/increment-counter")
-    public ResponseEntity<String> incrementPostCounter(@PathVariable Integer id) {
+    public ResponseEntity<String> incrementPostCounter(@PathVariable Integer id, BindingResult bindingResult) {
+
+        if (bindingResult.hasErrors()) {
+            throw new BadRequestException(bindingResult.getAllErrors());
+        }
+
         postService.increaseSharedCounter(id);
         return ResponseEntity.ok().body("Shared Counter of the post was increased.");
     }
